abstract masking to a new class
